% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hbar_fill.R
\name{geom_hbar_fill}
\alias{geom_hbar_fill}
\title{Draw A Horizontal Bar Chart with Fill Column}
\usage{
geom_hbar_fill(
  data,
  var_main,
  var_fill,
  wt = NULL,
  percent = FALSE,
  title = NULL,
  dec_places = 2,
  xlab = NULL,
  ylab = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{A data frame with variables \code{var_main} and \code{var_fill}}

\item{var_main}{The main variable to go on the x-axis}

\item{var_fill}{The fill variable to go on the y-axis in the bars}

\item{wt}{The weights for tallying}

\item{percent}{Plot the percentages}

\item{title}{The title of the chart}

\item{dec_places}{Rounding decimal places}

\item{xlab}{The label of the x-axis}

\item{ylab}{The label of the y-axis}

\item{caption}{Optional caption for information}
}
\value{
A filled (2d) horizontal bar chart
}
\description{
Draw A Horizontal Bar Chart with Fill Column
}
\examples{
library(ggplot2)
geom_hbar_fill(mpg, "class", "drv")
}
