% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vbar.R
\name{geom_vbar}
\alias{geom_vbar}
\title{Draw a vertical bar chart}
\usage{
geom_vbar(
  data,
  target,
  wt = NULL,
  percent = FALSE,
  percent_format = FALSE,
  title = NULL,
  dec_places = 1,
  labels_width = 0.9 * getOption("width"),
  monochrome = TRUE,
  xlab = NULL,
  ylab = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{Data set with column named \code{target}}

\item{target}{The column to put on the bar chart}

\item{wt}{The weights for tallying}

\item{percent}{Plot the percentages}

\item{percent_format}{Format label as percentage}

\item{title}{Title of the bar chart}

\item{dec_places}{Rounding decimal places}

\item{labels_width}{The number of characters per line of label text}

\item{monochrome}{Use one color}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{caption}{Caption to go below the x-axis}
}
\value{
The bar chart object
}
\description{
Draw a vertical bar chart
}
\examples{
df <- data.frame(x = sample(1:4, 20, replace = TRUE))
geom_vbar(df, "x")
}
