% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{order_factors_by_count}
\alias{order_factors_by_count}
\title{Order a Data Frame by Column Frequency Count}
\usage{
order_factors_by_count(data, target, wt = NULL)
}
\arguments{
\item{data}{A data frame with column named \code{target}}

\item{target}{A column to "factorize" and order by frequency of factors}

\item{wt}{Name of column with weights to calculate frequencies}
}
\value{
The data frame with \code{target} "factorized" and ordered
}
\description{
Order a Data Frame by Column Frequency Count
}
\examples{
df <- data.frame(x = sample(1:10, 20, replace = TRUE))
order_factors_by_count(df, "x")

}
